Deephaven AggOps

An AggOp is the return type of one of Deephaven's agg functions. It is a function that performs the computation specified by the agg function. These are intended to be passed directly to agg_by() or agg_all_by(), and should never be instantiated directly be user code. For more information, see the vignette on agg functions with vignette("agg_by").

If multiple tables have the same schema and the same aggregations need to be applied to each table, saving these objects directly in a variable may be useful to avoid having to re-create them each time:

aggregations <- c(agg_min("XMin = X", "YMin = Y"),
                  agg_max("XMax = X", "YMax = Y"))

result1 <- th1$agg_by(aggregations, by="Group")
result2 <- th2$agg_by(aggregations, by="Group")

In this example, aggregations would be a vector of two AggOps that can be reused in multiple calls to agg_by().

Methods

Public methods

AggOp$new()

Method new()

Usage

AggOp$new(aggregation, agg_name, ...)