Deephaven OperationControl

An OperationControl is the return type of Deephaven's op_control function. It is a function that determines how special values will be handled in the context of an update_by operation. An OperationControl is intended to be passed directly to a uby function, and should never be instantiated directly be user code.

If you plan to use the same operation control parameters for multiple uby functions in one or more update_by calls, consider creating an OperationControl object and reusing it. For example:

opc <- op_control(on_null = 'skip', on_nan = 'poison', big_value_context = 'decimal64')

result <- th$update_by(c(uby_ema_tick(5, c("XEma = X", "YEma = Y"), opc),
                         uby_emstd_tick(5, c("XEmstd = X", "YEmstd = Y"), opc)),
                       by="Group")

Methods

Public methods

OperationControl$new()

Method new()

Usage

OperationControl$new(op_control)